refactor: dependency add logic
Extract imperative non-cli logic to own function. This way it can be tested in integration tests without the cli. Actual functionality was only changed minimally.
