+++ src/cli/cmd-add.ts
@@ -1,78 +1,21 @@
 import { Logger } from "npmlog";
-import { CustomError } from "ts-custom-error";
-import { Err } from "ts-results-es";
-import { PackumentNotFoundError } from "../common-errors";
-import { NodeType, traverseDependencyGraph } from "../domain/dependency-graph";
-import { DomainName } from "../domain/domain-name";
-import { compareEditorVersion, EditorVersion } from "../domain/editor-version";
-import { tryGetTargetEditorVersionFor } from "../domain/package-manifest";
+import { addDependenciesUsing } from "../app/add-dependencies";
+import { determineEditorVersionUsing } from "../app/determine-editor-version";
+import { loadRegistryAuthUsing } from "../app/get-registry-auth";
 import {
   makePackageReference,
   PackageReference,
-  splitPackageReference,
 } from "../domain/package-reference";
-import { PackageUrl } from "../domain/package-url";
-import {
-  addTestable,
-  hasDependency,
-  mapScopedRegistry,
-  setDependency,
-  UnityProjectManifest,
-} from "../domain/project-manifest";
-import { unityRegistryUrl } from "../domain/registry-url";
-import {
-  addScope,
-  makeEmptyScopedRegistryFor,
-} from "../domain/scoped-registry";
-import { SemanticVersion } from "../domain/semantic-version";
-import {
-  loadProjectManifestUsing,
-  saveProjectManifestUsing,
-} from "../io/project-manifest-io";
-import { DebugLog } from "../logging";
-import { areArraysEqual } from "../utils/array-utils";
-import {
-  logFailedDependency,
-  logResolvedDependency,
-} from "./dependency-logging";
-import { CmdOptions } from "./options";
-import { ResultCodes } from "./result-codes";
-
-import { resolveDependenciesUsing } from "../app/dependency-resolving";
-import { determineEditorVersionUsing } from "../app/determine-editor-version";
-import { loadRegistryAuthUsing } from "../app/get-registry-auth";
-import { FetchRegistryPackumentVersion } from "../app/get-registry-packument-version";
-import { ResolvePackumentVersionError } from "../domain/packument";
-import { unityRegistry } from "../domain/registry";
 import { getUserUpmConfigPathFor } from "../domain/upm-config";
 import type { CheckUrlExists } from "../io/check-url";
 import type { GetRegistryPackument } from "../io/packument-io";
 import { getHomePathFromEnv } from "../io/special-paths";
 import type { ReadTextFile, WriteTextFile } from "../io/text-file-io";
-import { partialApply } from "../utils/fp-utils";
-import { isZod } from "../utils/zod-utils";
+import { DebugLog } from "../logging";
+import { recordEntries } from "../utils/record-utils";
+import { CmdOptions } from "./options";
 import { parseEnvUsing } from "./parse-env";
-
-export class PackageIncompatibleError extends CustomError {
-  constructor(
-    public readonly packageRef: PackageReference,
-    public readonly editorVersion: EditorVersion
-  ) {
-    super();
-  }
-}
-
-export class UnresolvedDependenciesError extends CustomError {
-  constructor(public readonly packageRef: PackageReference) {
-    super();
-  }
-}
-
-export class CompatibilityCheckFailedError extends CustomError {
-  constructor(public readonly packageRef: PackageReference) {
-    super();
-  }
-}
+import { ResultCodes } from "./result-codes";
 
 /**
  * Options passed to the add command.
@@ -104,24 +47,6 @@ type AddCmd = (
   options: AddOptions
 ) => Promise<AddResultCode>;
 
-function pickMostFixable(
-  a: Err<ResolvePackumentVersionError>,
-  b: Err<ResolvePackumentVersionError>
-): Err<ResolvePackumentVersionError> {
-  // Anything is more fixable than packument-not-found
-  if (
-    a.error instanceof PackumentNotFoundError &&
-    !(b.error instanceof PackumentNotFoundError)
-  )
-    return b;
-  else if (
-    b.error instanceof PackumentNotFoundError &&
-    !(a.error instanceof PackumentNotFoundError)
-  )
-    return a;
-  return a;
-}
-
 /**
  * Makes a {@link AddCmd} function.
  */
@@ -133,22 +58,6 @@ export function makeAddCmd(
   log: Logger,
   debugLog: DebugLog
 ): AddCmd {
-  const loadProjectManifest = partialApply(
-    loadProjectManifestUsing,
-    readTextFile,
-    debugLog
-  );
-
-  const saveProjectManifest = partialApply(
-    saveProjectManifestUsing,
-    writeTextFile
-  );
-
-  const getRegistryPackumentVersion = partialApply(
-    FetchRegistryPackumentVersion,
-    fetchPackument
-  );
-
   return async (pkgs, options) => {
     if (!Array.isArray(pkgs)) pkgs = [pkgs];
 
@@ -166,6 +75,9 @@ export function makeAddCmd(
         "editor.version",
         `${editorVersion} is unknown, the editor version check is disabled`
       );
+
+    const projectDirectory = env.cwd;
+
     const homePath = getHomePathFromEnv(process.env);
     const upmConfigPath = getUserUpmConfigPathFor(
       process.env,
@@ -180,204 +92,39 @@ export function makeAddCmd(
       env.primaryRegistryUrl
     );
 
-    const tryAddToManifest = async function (
-      manifest: UnityProjectManifest,
-      pkg: PackageReference
-    ): Promise<[UnityProjectManifest, boolean]> {
-      // is upstream package flag
-      let isUpstreamPackage = false;
-      // parse name
-      const [name, requestedVersion] = splitPackageReference(pkg);
+    const addResults = await addDependenciesUsing(
+      readTextFile,
+      writeTextFile,
+      fetchPackument,
+      checkUrlExists,
+      debugLog,
+      projectDirectory,
+      typeof editorVersion === "string" ? null : editorVersion,
+      primaryRegistry,
+      env.upstream,
+      options.force === true,
+      options.test === true,
+      pkgs
+    );
 
-      // packages that added to scope registry
-      const pkgsInScope = Array.of<DomainName>();
-      let versionToAdd = requestedVersion;
-      if (
-        requestedVersion === undefined ||
-        !isZod(requestedVersion, PackageUrl)
-      ) {
-        let resolveResult = await getRegistryPackumentVersion(
-          name,
-          requestedVersion,
-          primaryRegistry
-        ).promise;
-        if (resolveResult.isErr() && env.upstream) {
-          const upstreamResult = await getRegistryPackumentVersion(
-            name,
-            requestedVersion,
-            unityRegistry
-          ).promise;
-          if (upstreamResult.isOk()) {
-            resolveResult = upstreamResult;
-            isUpstreamPackage = true;
-          } else {
-            resolveResult = pickMostFixable(resolveResult, upstreamResult);
-          }
+    recordEntries(addResults)
+      .map(([packageName, addResult]) => {
+        switch (addResult.type) {
+          case "added":
+            return `added ${makePackageReference(
+              packageName,
+              addResult.version
+            )}`;
+          case "upgraded":
+            return `modified ${packageName} ${addResult.fromVersion} => ${addResult.toVersion}`;
+          case "noChange":
+            return `existed ${makePackageReference(
+              packageName,
+              addResult.version
+            )}`;
         }
-
-        if (resolveResult.isErr()) throw resolveResult.error;
-
-        const packumentVersion = resolveResult.value.packumentVersion;
-        versionToAdd = packumentVersion.version;
-
-        // Only do compatibility check when we have a valid editor version
-        if (typeof editorVersion !== "string") {
-          let targetEditorVersion: EditorVersion | null;
-          try {
-            targetEditorVersion =
-              tryGetTargetEditorVersionFor(packumentVersion);
-          } catch (error) {
-            if (!options.force) {
-              const packageRef = makePackageReference(name, versionToAdd);
-              debugLog(
-                `"${packageRef}" is malformed. Target editor version could not be determined.`
-              );
-              throw new CompatibilityCheckFailedError(packageRef);
-            }
-            targetEditorVersion = null;
-          }
-
-          // verify editor version
-          const isCompatible =
-            targetEditorVersion === null ||
-            compareEditorVersion(editorVersion, targetEditorVersion) >= 0;
-          if (!isCompatible && !options.force)
-            throw new PackageIncompatibleError(
-              makePackageReference(name, versionToAdd),
-              targetEditorVersion!
-            );
-        }
-
-        // pkgsInScope
-        if (!isUpstreamPackage) {
-          debugLog(`fetch: ${makePackageReference(name, requestedVersion)}`);
-          const dependencyGraph = await resolveDependenciesUsing(
-            checkUrlExists,
-            fetchPackument,
-            [primaryRegistry, unityRegistry],
-            name,
-            versionToAdd,
-            true
-          );
-
-          let isAnyDependencyUnresolved = false;
-          for (const [
-            dependencyName,
-            dependencyVersion,
-            dependency,
-          ] of traverseDependencyGraph(dependencyGraph)) {
-            if (dependency.type === NodeType.Failed) {
-              logFailedDependency(
-                log,
-                dependencyName,
-                dependencyVersion,
-                dependency
-              );
-              // If the manifest already has the dependency than it does not
-              // really matter that it was not resolved.
-              if (!hasDependency(manifest, dependencyName))
-                isAnyDependencyUnresolved = true;
-              continue;
-            }
-            if (dependency.type === NodeType.Unresolved) continue;
-
-            const dependencyRef = makePackageReference(
-              dependencyName,
-              dependencyVersion
-            );
-            logResolvedDependency(debugLog, dependencyRef, dependency.source);
-
-            const isUnityPackage =
-              dependency.source === "built-in" ||
-              dependency.source === unityRegistryUrl;
-            if (isUnityPackage) continue;
-
-            // add depsValid to pkgsInScope.
-            pkgsInScope.push(dependencyName);
-          }
-
-          // print suggestion for depsInvalid
-
-          if (isAnyDependencyUnresolved && !options.force)
-            throw new UnresolvedDependenciesError(
-              makePackageReference(name, versionToAdd)
-            );
-        } else pkgsInScope.push(name);
-      }
-      // add to dependencies
-      const oldVersion = manifest.dependencies[name];
-      // Whether a change was made that requires overwriting the manifest
-      let dirty = false;
-      manifest = setDependency(
-        manifest,
-        name,
-        versionToAdd as PackageUrl | SemanticVersion
-      );
-      if (!oldVersion) {
-        // Log the added package
-        log.notice(
-          "manifest",
-          `added ${makePackageReference(name, versionToAdd)}`
-        );
-        dirty = true;
-      } else if (oldVersion !== versionToAdd) {
-        // Log the modified package version
-        log.notice(
-          "manifest",
-          `modified ${name} ${oldVersion} => ${versionToAdd}`
-        );
-        dirty = true;
-      } else {
-        // Log the existed package
-        log.notice(
-          "manifest",
-          `existed ${makePackageReference(name, versionToAdd)}`
-        );
-      }
-
-      if (!isUpstreamPackage && pkgsInScope.length > 0) {
-        manifest = mapScopedRegistry(
-          manifest,
-          primaryRegistry.url,
-          (initial) => {
-            let updated =
-              initial ?? makeEmptyScopedRegistryFor(primaryRegistry.url);
-
-            updated = pkgsInScope.reduce(addScope, updated!);
-            dirty =
-              !areArraysEqual(updated!.scopes, initial?.scopes ?? []) || dirty;
-
-            return updated;
-          }
-        );
-      }
-      if (options.test) manifest = addTestable(manifest, name);
-
-      return [manifest, dirty];
-    };
-
-    // load manifest
-    let manifest = await loadProjectManifest(env.cwd);
-
-    // add
-    let dirty = false;
-    for (const pkg of pkgs) {
-      const [newManifest, manifestChanged] = await tryAddToManifest(
-        manifest,
-        pkg
-      );
-      if (manifestChanged) {
-        manifest = newManifest;
-        dirty = true;
-      }
-    }
-
-    // Save manifest
-    if (dirty) {
-      await saveProjectManifest(env.cwd, manifest);
-      // print manifest notice
-      log.notice("", "please open Unity project to apply changes");
-    }
+      })
+      .forEach((message) => log.info("", message));
 
     return ResultCodes.Ok;
   };
